Adds single image posting FE&BE
Change FE createPost to use forms instead of JSON. Allows to attach binary data in fields :)

@DerekLeeCS

I change a few things about ur functions--

createImage and other image creating functions that need form field names like "file" and "fileType" are now variables

When we do multi-image posts(if we do), I think FE should send a "numberOfImages" form field and then each image is put under form fields "image0" "image1" "image2" ... "imageType0" "imageType1" "ImageType3" and so on

BE will be able to loop through and create all the images and store the imageID's of all of them in a list ( using the Image returned from createImage() ), and then call postStore with all the information.

As such createImage() is now private and returns Image.